% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ch6ex}
\alias{ch6ex}
\title{Dataset for Chapter 6 Exercises, An R Companion to Political Analysis, Third Edition}
\format{
A data frame with 100 rows and 7 variables.
\describe{
  \item{id}{Unique identification number for each subject}
  \item{subject_name}{Last name of subject}
  \item{age}{Age of subject in years}
  \item{partyid}{Self-identified party identification from pretest}
  \item{prior_classes}{Number of political science classes subject has completed}
  \item{site_viewed}{Did subject view the candidate's old or new site in the experiment?}
  \item{site_rating}{Subject's rating of candidate on feeling thermometer after viewing web site}
}
}
\source{
Barry C. Edwards and Philip H. Pollock, III
}
\usage{
ch6ex
}
\description{
A hypothetical dataset of human subject research
}
\keyword{internal}
