% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{debate}
\alias{debate}
\title{Debate Experiment dataset for R Companion to Political Analysis, Third Edition}
\format{
A data frame with 171 rows and 14 variables.
\describe{
  \item{obs}{Unique identification number for each subject}
  \item{assignment}{Name of condition subject was assigned to}
  \item{tv}{Did subject watch debate on TV? 1 = yes, 0 = no}
  \item{debinfo}{Number correct answers on five question quiz about the debate.}
  \item{catholic}{Is subject Catholic? 1 = yes, 0 = no}
  \item{issues}{Which candidate do you agree with on policy issues?}
  \item{integrity}{Which candidate has more integrity?}
  \item{leadership}{Which candidate is more effective leader?}
  \item{empathy}{Which candidate has more empathy?}
  \item{sophdum}{Is respondent a sophomore? 1 = yes, 0 = no.}
  \item{won}{Which candidate won the debate? 1 = Kennedy ... 4 = Tie ... 7 = Nixon}
  \item{pid}{Self-reported partisan identification on standard 1-7 scale}
  \item{ideology}{Self-reported political ideology on standard 1-7 scale}
  \item{gender}{Subject's gender, 0 = male, 1 = female.}
}
}
\source{
Jamie Druckman.  See Appendix of printed textbook for further information.
}
\usage{
debate
}
\description{
A dataset with variables about students who participated in an experiment. This dataset is used to demonstrate application of R to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}
