% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.info.functions.r
\name{graph.info.functions}
\alias{graph.info.functions}
\alias{Colors}
\alias{lineType}
\alias{plotChar}
\title{Reference information on graphics parameters, like colors, line types, plotting characters}
\usage{
Colors()

lineType()

plotChar()
}
\value{
No value returned

No value returned

No return
}
\description{
Students can use these functions as quick references on some of R's plotting options.

The \strong{Colors} function produces plot of colors available in R. Colored cells labels with numeric value that corresponds to that color. No values supplied to function.

The \strong{lineType} function displays line types for R plots. No values supplied to function.

The \strong{plotChar} function displays plotting characters available in R. No values supplied to function.
}
\examples{
  Colors()
  lineType()
  plotChar()
}
\keyword{internal}
