% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{inverse.logit}
\alias{inverse.logit}
\title{Converts linear predictor to probability}
\usage{
inverse.logit(logged_odds)
}
\arguments{
\item{logged_odds}{A numeric value, or vector of numeric values.}
}
\value{
Returns predicted probability corresponding the the logged odds value, a numeric value between 0 and 1.
}
\description{
Calculates predicted probability for a given logged odds value. This function calculates predicted probability for a given logged odds value, often useful for plotting or reporting predicted probabilities.
}
\examples{
   library(RCPA3)
   
   inverse.logit(0)
   inverse.logit(-5:5)
}
\keyword{internal}
