% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.thru.functions.r
\name{pass.thru.functions}
\alias{pass.thru.functions}
\alias{scatterplot}
\alias{svyby}
\alias{svydesign}
\alias{svyplot}
\title{Imported Functions Made Available to Package Users}
\usage{
scatterplot(...)

svyby(...)

svydesign(...)

svyplot(...)
}
\arguments{
\item{...}{See source package help files for details on function usage.}
}
\value{
A svyby class object, see \code{\link[survey]{svyby}} documentation for details.

A survey.design class object, see \code{\link[survey]{svydesign}} documentation for details.
}
\description{
A number of functions we used in our \emph{R Companion to Political Analysis, 2nd Edition} are not used in \emph{R Companion to Political Analysis, 3rd Edition}.
For backwards compatibility, we have some functions that simply pass through arguments to a function in another package (allowing user to call them without loading additional packages).

The \code{\link[car]{scatterplot}} function is imported from the car package.
See \code{\link[car]{scatterplot}} documentation for details.

The \code{\link[survey]{svyby}} function is imported from the survey package.
See \code{\link[survey]{svyby}} documentation for details.

The \code{\link[survey]{svydesign}} function is imported from the survey package.
See \code{\link[survey]{svydesign}} documentation for details.

The \code{\link[survey]{svyplot}} function is imported from the survey package.
See \code{\link[survey]{svyplot}} documentation for details.
}
\keyword{internal}
