% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.names.r
\name{random.names}
\alias{random.names}
\title{Generates random sample of nationally representative surnames}
\usage{
random.names(n)
}
\arguments{
\item{n}{The number of random surnames to be generated}
}
\value{
A vector of surnames
}
\description{
Generates a random sample of surnames of any length. Surnames and probabilities from Census Bureau, see <https://www.census.gov/topics/population/genealogy/data/2010_surnames.html>. Used to illustrate random sampling methods.
}
\examples{
   library(RCPA3)
   
   random.names(10)
}
\keyword{internal}
