% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetFactory.r
\name{widgetFactory}
\alias{widgetFactory}
\title{An interactive game to practice using R functions.}
\usage{
widgetFactory()
}
\value{
No value returned
}
\description{
A fun, interactive game to practice using R functions. Players must execute functions to make widgets per strict specifications to win the game. The Widget Factory needs your help!
}
\section{RCPA3 Package Tutorial Videos}{

\itemize{
\item \href{https://www.youtube.com/playlist?list=PL3jY4WDTUxoNqrxSSQH4q7XPLPYipeNCu}{Complete Playlist of RCPA3 Package Tutorial Videos}, includes video for this function and many more.
}
}

\section{Textbook Reference}{

\itemize{
\item Philip H. Pollock and Barry C. Edwards, \emph{An R Companion to Political Analysis, 3rd Edition} (Thousand Oaks, CA: Sage Publications, Forthcoming 2022), Chapter 1.
}
}

\section{Online Resource}{

\itemize{
\item \href{https://www.poliscidata.com/pages/rDemosResources.php?chapter=1}{Tutorials & Resources for Using R for Data Analysis}, compiled by Barry C. Edwards
}
}

\examples{
 
library(RCPA3)
 
# Play the Widget Factory game
\dontrun{
widgetFactory()
}
}
