% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{wtd.median}
\alias{wtd.median}
\title{Find Median of Variable with Option to Weight Observations}
\usage{
wtd.median(x, w = NULL, data, ...)
}
\arguments{
\item{x}{A variable}

\item{w}{(Optional) Sampling weights of variable}

\item{data}{(Optional) Name of dataset that contains x (and w) variable}
}
\value{
Returns the median value of the variable
}
\description{
Takes in variable and finds median, works with sampling weights. Makes use of the wtd.quantile function, part of the Hmisc package.
}
\examples{
   library(RCPA3)
   
   wtd.median(x=nes$approve.pres.hc, w=nes$wt)
   wtd.median(x=nes$approve.pres.hc)
}
\keyword{internal}
