% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{wtd.skewness}
\alias{wtd.skewness}
\title{Calculates Skewness of Numeric Variable with Option to Weight Observations}
\usage{
wtd.skewness(x, w, data, digits = 3)
}
\arguments{
\item{x}{Variable (must be a numeric vector)}

\item{w}{(Optional) Weights variable}

\item{data}{(Optional) Dataset}

\item{digits}{(Optional) Number of digits after decimal point (default: 3)}
}
\value{
Returns the skewness of the variable, a numeric value.
}
\description{
Calculates Skewness of Numeric Variable with Option to Weight Observations
}
\examples{
   library(RCPA3)
   
   \donttest{
   wtd.skewness(x=nes$ft.socialists, w=nes$wt)
   wtd.skewness(x=nes$ft.socialists)
   }
}
\keyword{internal}
